+++ roqoqo/tests/integration/quantum_program.rs
@@ -24,6 +24,8 @@ use roqoqo::QuantumProgram;
 #[cfg(feature = "json_schema")]
 use schemars::schema_for;
 use std::collections::HashMap;
+#[cfg(feature = "unstable_parallel_run")]
+use std::time::Instant;
 
 #[derive(Debug, Clone, Copy)]
 struct TestBackend;
@@ -195,6 +197,57 @@ fn test_registers() {
     let result_fail = program.run_registers(backend, &[0.0]);
     assert!(result_fail.is_err());
     let result_fail = program.run_registers(backend, &[0.0, 1.0, 3.0]);
+    assert!(result_fail.is_err());
+}
+
+#[cfg(feature = "unstable_parallel_run")]
+#[test]
+fn test_parallel() {
+    // setting ub BR measurement
+    let bri = CheatedInput::new(2);
+    let mut circs: Vec<Circuit> = Vec::new();
+    let mut circ1 = Circuit::new();
+    let mut circ1_subs = Circuit::new();
+    circ1 += operations::RotateX::new(0, "theta".into());
+    circ1_subs += operations::RotateX::new(0, 0.0.into());
+    let mut circ2 = Circuit::new();
+    let mut circ2_subs = Circuit::new();
+    circ2 += operations::RotateZ::new(0, "theta2".into());
+    circ2_subs += operations::RotateZ::new(0, 1.0.into());
+    circs.push(circ1);
+    let br = Cheated {
+        constant_circuit: Some(circ2),
+        circuits: circs.clone(),
+        input: bri,
+    };
+
+    let input_parameter_names = vec!["theta".to_string(), "theta2".to_string()];
+    let program = QuantumProgram::Cheated {
+        measurement: br,
+        input_parameter_names,
+    };
+
+    let backend = TestBackend;
+
+    let mut now = Instant::now();
+    let result = program.run_parallel(backend, &[vec![0.0, 1.0]]);
+    let one_parameter_set_time = now.elapsed();
+    assert!(result.is_ok());
+
+    let number_parameter_set = 1000;
+    now = Instant::now();
+    let result = program.run_parallel(backend, &vec![vec![0.0, 1.0]; number_parameter_set]);
+    let multiple_parameter_set_time = now.elapsed();
+    assert!(result.is_ok());
+    assert!(
+        multiple_parameter_set_time
+            < one_parameter_set_time.mul_f32(0.9) * number_parameter_set as u32
+    );
+
+    let result_fail = program.run_parallel(backend, &[vec![0.0]]);
+    assert!(result_fail.is_err());
+
+    let result_fail = program.run_parallel(backend, &[vec![0.0, 1.0, 3.0]]);
     assert!(result_fail.is_err());
 }