+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ## Unreleased
 
+* Added the `run_parallel` method to QuantumProgram under the `unstable_parallel_run` feature.
+
 ## 1.20.0
 
 * Added the `number_of_qubits` method to the `Circuit` class.